# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSmart API",
    "description": "\n## 🏥 MediSmart — Hệ thống tra cứu thuốc & tư vấn sức khỏe thông minh\n\n### 🔑 Authentication\n- Lấy token: `POST /api/v1/auth/login`\n- Gắn vào header: `Authorization: Bearer {access_token}`\n- Access token hết hạn sau **30 phút** → dùng `POST /api/v1/auth/refresh`\n- Logout → token bị thu hồi ngay lập tức (Redis blacklist)\n\n### 👥 Phân quyền\n| Role  | Quyền truy cập |\n|-------|----------------|\n| User  | Hồ sơ cá nhân, tra cứu thuốc, chatbot, lịch sử |\n| Admin | Tất cả của User + quản lý dữ liệu + system logs |\n\n### 📄 Response Format\nTất cả responses đều theo format:\n```json\n{\n  \"success\": true,\n  \"message\": \"Thành công\",\n  \"data\": { ... }\n}\n```\n\n### ⚠️ Rate Limiting\n- Login: 5 lần / 15 phút / IP\n- Gửi lại email xác thực: 1 lần / 5 phút / user\n- Reset mật khẩu: 1 lần / 5 phút / user\n    ",
    "contact": {
      "name": "MediSmart Support",
      "email": "support@medismart.vn"
    },
    "license": {
      "name": "Private"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Lấy thông tin cá nhân",
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Cập nhật thông tin cá nhân",
        "operationId": "update_profile_api_v1_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/password": {
      "put": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Đổi mật khẩu",
        "operationId": "change_password_api_v1_users_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "post": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Upload ảnh đại diện",
        "description": "Chấp nhận JPEG, PNG, WebP. Tối đa 2MB.",
        "operationId": "upload_avatar_api_v1_users_me_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_users_me_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/prescriptions": {
      "get": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Danh sách đơn thuốc",
        "description": "Hỗ trợ tìm kiếm theo tên và lọc theo trạng thái.",
        "operationId": "list_prescriptions_api_v1_users_me_prescriptions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tìm theo tên đơn thuốc",
              "title": "Search"
            },
            "description": "Tìm theo tên đơn thuốc"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc: active | completed",
              "title": "Status"
            },
            "description": "Lọc: active | completed"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Tạo đơn thuốc mới",
        "operationId": "create_prescription_api_v1_users_me_prescriptions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Xóa nhiều đơn thuốc",
        "operationId": "delete_many_prescriptions_api_v1_users_me_prescriptions_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/prescriptions/{prescription_id}": {
      "get": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Chi tiết đơn thuốc",
        "operationId": "get_prescription_api_v1_users_me_prescriptions__prescription_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Cập nhật đơn thuốc",
        "operationId": "update_prescription_api_v1_users_me_prescriptions__prescription_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prescription Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Xóa đơn thuốc",
        "description": "Chỉ xóa được đơn có trạng thái 'completed'.",
        "operationId": "delete_prescription_api_v1_users_me_prescriptions__prescription_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/prescriptions/{prescription_id}/interactions": {
      "get": {
        "tags": [
          "👤 User Profile"
        ],
        "summary": "Kiểm tra tương tác thuốc",
        "description": "Kiểm tra tương tác giữa các thuốc trong đơn (cần ít nhất 2 thuốc có drug_id).",
        "operationId": "check_interactions_api_v1_users_me_prescriptions__prescription_id__interactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "type": "string",
        "enum": [
          "local",
          "google"
        ],
        "title": "AuthProvider"
      },
      "Body_upload_avatar_api_v1_users_me_avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_v1_users_me_avatar_post"
      },
      "BulkDeleteRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "BulkDeleteRequest"
      },
      "ChangePasswordRequest": {
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          },
          "confirm_password": {
            "type": "string",
            "title": "Confirm Password"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password",
          "confirm_password"
        ],
        "title": "ChangePasswordRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrescriptionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrescriptionStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": "active"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemCreate"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "name",
          "items"
        ],
        "title": "PrescriptionCreate"
      },
...
      "PrescriptionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/PrescriptionStatus"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "drug_count": {
            "type": "integer",
            "title": "Drug Count",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "status",
          "items",
          "created_at",
          "updated_at",
          "drug_count"
        ],
        "title": "PrescriptionResponse",
        "description": "Chi tiết đầy đủ bao gồm items; drug_count computed từ items."
      },
      "PrescriptionStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed"
        ],
        "title": "PrescriptionStatus"
      },
      "PrescriptionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrescriptionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PrescriptionItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "title": "PrescriptionUpdate"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "auth_provider": {
            "$ref": "#/components/schemas/AuthProvider"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "role",
          "auth_provider",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}